vim: upgrade to 8.2 with huge features and dynamic python3 (#4030)
Configuring `vim` with dynamic Python 3 support means it will attempt to
load `libpython` using `dlopen`, allowing it to gracefully fail if it is
unavailable. This removes the need for a dependency on the `python3`
SPK and avoids shipping of a redundant set of Python 3 bins/libs within
the `vim` SPK package.

`vimtutor` is a script, not a binary, so cannot be stripped.